+++ packages/env/index.ts
@@ -1,8 +1,8 @@
-import { create } from '@zazuko/env-node'
-import type { DefaultEnv, DerivedEnvironment } from '@zazuko/env'
+import BaseEnv, { Environment as ZazukoEnv } from '@zazuko/env-node'
 import NamespacesFactory from './lib/Namespaces.js'
 import ConstantsFactory from './lib/Constants.js'
 
-export type Environment = DerivedEnvironment<DefaultEnv, ConstantsFactory>
+const env = new ZazukoEnv([NamespacesFactory, ConstantsFactory], { parent: BaseEnv })
 
-export default create<typeof NamespacesFactory | typeof ConstantsFactory>(NamespacesFactory, ConstantsFactory)
+export default env
+export type Environment = typeof env