+++ packages/rdf/lib/metadata/applyOptions.js
@@ -1,20 +1,32 @@
 import rdf from 'barnard59-env'
-import fromStream from 'rdf-dataset-ext/fromStream.js'
-import * as ns from '../namespaces.js'
-import { xsd } from '../namespaces.js'
 import { wellKnownDatasetClasses, wellKnownDatasetClassesWithDcterms } from './datasetClasses.js'
 import { namedDateLiterals } from './namedDateLiterals.js'
 
+/**
+ * @param {import('@rdfjs/types').DatasetCore} dataset
+ * @param {Set<import('@rdfjs/types').Term>} classes
+ * @return {Set<import('@rdfjs/types').Quad_Subject>}
+ */
 function subjectsWithDatasetType(dataset, classes) {
+  /**
+   * @type {Set<import('@rdfjs/types').Quad_Subject>}
+   */
   const result = rdf.termSet()
   ;[...dataset]
-    .filter(quad => (quad.predicate.equals(ns.rdf.type) && classes.has(quad.object)))
+    .filter(quad => (quad.predicate.equals(rdf.ns.rdf.type) && classes.has(quad.object)))
     .forEach(quad => {
       result.add(quad.subject)
     })
   return result
 }
 
+/**
+ * @param {import('@rdfjs/types').DatasetCore} dataset
+ * @param {Set<import('@rdfjs/types').Term>} datasetClasses
+ * @param {import('@rdfjs/types').Quad_Predicate} predicate
+ * @param {import('@rdfjs/types').Quad_Object} object
+ * @return {import('@rdfjs/types').DatasetCore}
+ */
 function updateOrInsert(dataset, datasetClasses, predicate, object) {
   const targetSubjects = subjectsWithDatasetType(dataset, datasetClasses)
 
@@ -31,34 +43,58 @@ function updateOrInsert(dataset, datasetClasses, predicate, object) {
   return dataset
 }
 
-function toDateLiteral(item) {
-  return typeof item === 'string' ? rdf.literal(item, xsd.dateTime) : item
-}
-
+/**
+ * @param {string | import('@rdfjs/types').NamedNode | undefined} item
+ * @return {import('@rdfjs/types').NamedNode | undefined}
+ */
 function toNamedNode(item) {
   return typeof item === 'string' ? rdf.namedNode(item) : item
 }
 
+/**
+ * @param {import('@rdfjs/types').Literal | import('./namedDateLiterals.js').NamedDateLiteral} value
+ * @param {import('./namedDateLiterals.js').Metadata} metadata
+ * @return {import('@rdfjs/types').Literal}
+ */
 function resolveNamedDate(value, metadata) {
-  return namedDateLiterals.has(value) ? namedDateLiterals.get(value)(metadata) : toDateLiteral(value)
+  if (typeof value !== 'string') {
+    return value
+  }
+
+  const factory = namedDateLiterals.get(value)
+
+  return factory ? factory(metadata) : rdf.literal(value, rdf.ns.xsd.dateTime)
 }
 
+/**
+ * @param {import('@rdfjs/types').Stream & import('stream').EventEmitter} quadStream
+ * @param {import('./namedDateLiterals.js').Metadata} [metadata]
+ * @param {{
+ *   dateModified?: import('./namedDateLiterals.js').NamedDateLiteral;
+ *   dateCreated?: import('./namedDateLiterals.js').NamedDateLiteral;
+ *   graph?: string | import('@rdfjs/types').NamedNode
+ * }} [options]
+ * @return {Promise<import('@rdfjs/types').DatasetCore>}
+ */
 async function applyOptions(quadStream, metadata = {}, options = {}) {
-  let dataset = await fromStream(rdf.dataset(), quadStream)
+  /**
+   * @type {import('@rdfjs/types').DatasetCore}
+   */
+  let dataset = await rdf.dataset().import(quadStream)
 
   // dateModified
   if (options.dateModified) {
     const dateModifiedLiteral = resolveNamedDate(options.dateModified, metadata)
 
-    dataset = updateOrInsert(dataset, wellKnownDatasetClassesWithDcterms, ns.dcterms.modified, dateModifiedLiteral)
-    dataset = updateOrInsert(dataset, wellKnownDatasetClasses, ns.schema.dateModified, dateModifiedLiteral)
+    dataset = updateOrInsert(dataset, wellKnownDatasetClassesWithDcterms, rdf.ns.dcterms.modified, dateModifiedLiteral)
+    dataset = updateOrInsert(dataset, wellKnownDatasetClasses, rdf.ns.schema.dateModified, dateModifiedLiteral)
   }
 
   // dateCreated
   if (options.dateCreated) {
     const dateCreatedLiteral = resolveNamedDate(options.dateCreated, metadata)
-    dataset = updateOrInsert(dataset, wellKnownDatasetClassesWithDcterms, ns.dcterms.created, dateCreatedLiteral)
-    dataset = updateOrInsert(dataset, wellKnownDatasetClasses, ns.schema.dateCreated, dateCreatedLiteral)
+    dataset = updateOrInsert(dataset, wellKnownDatasetClassesWithDcterms, rdf.ns.dcterms.created, dateCreatedLiteral)
+    dataset = updateOrInsert(dataset, wellKnownDatasetClasses, rdf.ns.schema.dateCreated, dateCreatedLiteral)
   }
 
   // Sets graph