+++ packages/rdf/lib/namespaces.js
@@ -1,5 +1,0 @@
-import $rdf from 'barnard59-env'
-
-const { cube, rdf, rdfs, sh, xsd, _void, dcat, schema, dcterms } = $rdf.ns
-
-export { cube, rdf, rdfs, sh, xsd, _void, dcat, schema, dcterms }