+++ packages/cube/CHANGELOG.md
@@ -1,5 +1,21 @@
 # barnard59-cube
 
+## 1.3.0
+
+### Minor Changes
+
+- f2e796c: Shape creation refactoring and improvements
+
+### Patch Changes
+
+- 7456a6a: added batch operation
+- Updated dependencies [82dbe7e]
+- Updated dependencies [c090ff2]
+- Updated dependencies [7456a6a]
+  - barnard59-rdf@3.4.0
+  - barnard59-shacl@1.3.1
+  - barnard59-base@2.4.0
+
 ## 1.2.0
 
 ### Minor Changes